Finished ! builtin, error handling, segfault patched

+++ src/open.c
@@ -6,7 +6,7 @@
 /*   By: agoublai <agoublai@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/16 12:36:27 by agirona           #+#    #+#             */
-/*   Updated: 2022/02/27 04:37:54 by agirona          ###   ########lyon.fr   */
+/*   Updated: 2022/02/28 01:27:07 by agirona          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,8 @@ int	verif_open_heredoc(void)
 	fd = open(".heredoc.tmp", O_WRONLY, 0644);
 	if (fd != -1)
 	{
-		close(fd);
+		if (close(fd) == -1)
+			return (return_perror(-3, "error ", errno));
 		return (unlink(".heredoc.tmp"));
 	}
 	return (1);
@@ -41,7 +42,8 @@ int	verif_open_in(t_cmd *cmd, char *fragment, int redir_type)
 		fd = open(cmd->redir_in, O_RDONLY, 0644);
 		if (fd == -1)
 			return (return_perror(-2, "error ", errno));
-		close(fd);
+		if (close(fd) == -1)
+			return (return_perror(-3, "error ", errno));
 	}
 	else if (verif_open_heredoc() < 0)
 		return (return_error(-3));
@@ -71,7 +73,8 @@ int	verif_open_out(t_cmd *cmd, char *fragment, int redir_type)
 		if (fd == -1)
 			return (return_perror(-2, "error ", errno));
 	}
-	close(fd);
+	if (close(fd) == -1)
+		return (return_perror(-3, "error ", errno));
 	return (1);
 }
 
@@ -85,10 +88,10 @@ int	verif_open(t_cmd *cmd, char *fragment, int redir_type)
 	if ((redir_type - 1) / 2 == 0)
 	{
 		ret = verif_open_in(cmd, fragment, redir_type);
-		if (ret == -1)
-			return (-5);
 		if (ret == -40)
 			return (-1);
+		if (ret != 1)
+			return (ret);
 	}
 	else
 	{
